Move `command` module into `connection` module
