+++ espflash/src/connection/reset.rs
@@ -10,12 +10,13 @@ use log::debug;
 use serialport::SerialPort;
 use strum::{Display, EnumIter, EnumString, VariantNames};
 
-use crate::{
+use super::{
     command::{Command, CommandType},
-    connection::{Connection, Port, USB_SERIAL_JTAG_PID},
-    error::Error,
-    flasher::FLASH_WRITE_SIZE,
+    Connection,
+    Port,
+    USB_SERIAL_JTAG_PID,
 };
+use crate::{error::Error, flasher::FLASH_WRITE_SIZE};
 
 /// Default time to wait before releasing the boot pin after a reset
 const DEFAULT_RESET_DELAY: u64 = 50; // ms
